netmon-backend/requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
bidict==0.21.2
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2020.11.15
pycodestyle==2.6.0
pylint==2.5.3
pymongo==3.11.3
python-engineio==4.1.0
python-socketio==5.1.0
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
